+++ common/shell.js
@@ -8414,6 +8414,9 @@ function resolveProviderMediaLink(
         streamShellYoutubeCleanupHideHomePromotions:
             "data-stream-shell-yt-hide-home-promotions",
 
+        streamShellYoutubeCleanupHidePlayables:
+            "data-stream-shell-yt-hide-playables",
+
         streamShellYoutubeCleanupHideVideoSidebar:
             "data-stream-shell-yt-hide-video-sidebar",
 
@@ -8540,6 +8543,9 @@ function resolveProviderMediaLink(
             false,
 
         streamShellYoutubeCleanupHideHomePromotions:
+            true,
+
+        streamShellYoutubeCleanupHidePlayables:
             true,
 
         streamShellYoutubeCleanupHideVideoSidebar: